Moved save functionality into the FaceEngineOutput smart component.
