+++ packages/veritone-widgets/src/widgets/FaceEngineOutput/styles.scss
@@ -1,3 +1,5 @@
+@import '../../styles/modules/variables';
+
 .editNewEntityDialogPaper {
   min-width: 640px;
   width: 640px;
@@ -65,4 +67,27 @@
 
 .faceDetectionDoneSnackbarContentText {
   font-size: 14px;
+}
+
+.actionButtonsEditMode {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+  align-items: center;
+  margin: 15px 80px 15px 0;
+  border-right: 1px solid $grey-4;
+  flex: none;
+
+  button {
+    height: 36px;
+    border-radius: 3px;
+    font-size: 14px;
+    font-weight: 500;
+    text-transform: uppercase;
+  }
+
+  .actionButtonEditMode {
+    width: 108px;
+    margin-right: 20px;
+  }
 }
